refactor(agents): extract AgentEditorMappers from AgentEditorViewModel
Move the pure load/save transforms (tool partitioning, applyAgentData,
model-parameter and handoff-edge codecs, tools-list assembly, and the
display-data mappers) out of the ViewModel companion into a top-level
AgentEditorMappers.kt. These are functions of their inputs only, so the
mapping is now unit-testable without a ViewModel, coroutines, or Android.

No behavior change.
